Add de Bruijn sequence algorithm for pattern generation and offsets
Replace the previous iterative pattern generation and offset functions
with the de Bruijn algorithm which creates a sequence in which every
possible length-n string on an alphabet of size k occurs exactly once
as a contiguous substring.

+++ src/args.rs
@@ -33,7 +33,7 @@ pub enum PatternSubcommand {
 #[derive(Debug, Args)]
 pub struct GeneratePattern {
     /// The length of the pattern
-    pub length: u32,
+    pub length: usize,
 }
 
 #[derive(Debug, Args)]
